@@ -1,6 +1,6 @@
 {
   "name": "@slate-collaborative/client",
-  "version": "0.0.3",
+  "version": "0.5.0",
   "files": [
     "lib"
   ],
@@ -24,14 +24,13 @@
     "watch": "yarn build:js -w"
   },
   "dependencies": {
+    "@babel/plugin-proposal-optional-chaining": "^7.9.0",
     "@babel/preset-react": "^7.0.0",
-    "@slate-collaborative/bridge": "^0.0.3",
-    "automerge": "^0.12.1",
-    "immutable": "^4.0.0-rc.12",
-    "react": "^16.9.0",
-    "slate": "^0.47.8",
-    "socket.io-client": "^2.2.0",
-    "typescript": "^3.6.3"
+    "@slate-collaborative/bridge": "^0.5.0",
+    "automerge": "^0.14.0",
+    "slate": "^0.57.2",
+    "socket.io-client": "^2.3.0",
+    "typescript": "^3.8.3"
   },
   "devDependencies": {
     "@babel/cli": "^7.6.0",
@@ -40,8 +39,7 @@
     "@babel/plugin-proposal-object-rest-spread": "^7.5.5",
     "@babel/preset-env": "^7.6.0",
     "@babel/preset-typescript": "^7.6.0",
-    "@types/react": "^16.9.2",
-    "@types/slate": "^0.47.1",
+    "@types/react": "^16.9.34",
     "@types/socket.io-client": "^1.4.32"
   },
   "directories": {
