@@ -1,14 +1,4 @@
-import defaultValue from './defaultValue'
-
-export const getClients = (io, nsp) =>
+export const getClients = (io: SocketIO.Server, nsp: string) =>
   new Promise((r, j) => {
-    io.of(nsp).clients((e, c) => (e ? j(e) : r(c)))
+    io.of(nsp).clients((e: any, c: any) => (e ? j(e) : r(c)))
   })
-
-export const defaultOptions = {
-  entry: 9000,
-  saveTreshold: 2000,
-  cursorAnnotationType: 'collaborative_selection'
-}
-
-export { defaultValue }
