@@ -1,20 +1,17 @@
-import { Operation, Operations, SyncDoc } from '../model'
+import { Operation } from 'slate'
 
 import node from './node'
-import mark from './mark'
 import text from './text'
-import annotation from './annotation'
 
-const setSelection = doc => doc
-const setValue = doc => doc
+import { SyncDoc } from '../model'
+import { toJS } from '../utils'
+
+const setSelection = (doc: any) => doc
 
 const opType = {
   ...text,
-  ...annotation,
   ...node,
-  ...mark,
   set_selection: setSelection
-  // set_value: setValue
 }
 
 const applyOperation = (doc: SyncDoc, op: Operation): SyncDoc => {
@@ -22,19 +19,19 @@ const applyOperation = (doc: SyncDoc, op: Operation): SyncDoc => {
     const applyOp = opType[op.type]
 
     if (!applyOp) {
-      console.log('operation', op.toJS())
       throw new TypeError(`Unsupported operation type: ${op.type}!`)
     }
 
-    return applyOp(doc, op)
+    return applyOp(doc, op as any)
   } catch (e) {
-    console.error(e)
+    console.error(e, op, toJS(doc))
 
     return doc
   }
 }
 
-const applySlateOps = (doc: SyncDoc, operations: Operations) =>
-  operations.reduce(applyOperation, doc)
+const applySlateOps = (doc: SyncDoc, operations: Operation[]): SyncDoc => {
+  return operations.reduce(applyOperation, doc)
+}
 
 export { applyOperation, applySlateOps }
