@@ -1,11 +1,12 @@
-import { SyncDoc } from '../model'
 import { InsertTextOperation, RemoveTextOperation } from 'slate'
+
 import { getTarget } from '../path'
+import { SyncDoc } from '../model'
 
 export const insertText = (doc: SyncDoc, op: InsertTextOperation): SyncDoc => {
   const node = getTarget(doc, op.path)
 
-  node.text.insertAt(op.offset, op.text)
+  node.text.insertAt(op.offset, ...op.text.split(''))
 
   return doc
 }
