@@ -1,4 +1,5 @@
 import * as Automerge from 'automerge'
+import { Node } from 'slate'
 
 import opInsert from './insert'
 import opRemove from './remove'
@@ -14,11 +15,11 @@ const byAction = {
 
 const rootKey = '00000000-0000-0000-0000-000000000000'
 
-const toSlateOp = (ops: Automerge.Diff[], doc) => {
-  const iterate = (acc, op) => {
+const toSlateOp = (ops: Automerge.Diff[], doc: Automerge.Doc<Node>) => {
+  const iterate = (acc: [any, any[]], op: Automerge.Diff): any => {
     const action = byAction[op.action]
 
-    const result = action ? action(op, acc) : acc
+    const result = action ? action(op, acc, doc) : acc
 
     return result
   }
