@@ -1,27 +1,28 @@
 import * as Automerge from 'automerge'
-import { TextJSON } from 'slate'
 
-export const createTextJSON = (text: string = ''): TextJSON => ({
-  object: 'text',
-  marks: [],
+import { toSync } from '../'
+
+import { Node } from 'slate'
+
+export const createText = (text: string = '') => ({
   text
 })
 
-export const createBlockJSON = (
+export const createNode = (
   type: string = 'paragraph',
-  text: string = ''
+  text: string = '',
+  data?: { [key: string]: any }
 ) => ({
-  object: 'block',
   type,
-  nodes: [createTextJSON(text)]
+  children: [createText(text)],
+  ...data
 })
 
-export const createValueJSON = () => ({
-  document: {
-    nodes: [createBlockJSON()]
-  }
+export const createValue = (children?: any): { children: Node[] } => ({
+  children: children || [createNode()]
 })
 
-export const createDoc = () => Automerge.from(createValueJSON())
+export const createDoc = (children?: any) =>
+  Automerge.from(toSync(createValue(children)))
 
-export const cloneDoc = doc => Automerge.change(doc, '', d => d)
+export const cloneDoc = (doc: any) => Automerge.change(doc, '', d => d)
