@@ -1,6 +1,16 @@
-const Connection = require('@slate-collaborative/backend')
-const defaultValue = require('./src/defaultValue')
+const { SocketIOConnection } = require('@slate-collaborative/backend')
 const express = require('express')
+
+const defaultValue = [
+  {
+    type: 'paragraph',
+    children: [
+      {
+        text: 'Hello collaborator!'
+      }
+    ]
+  }
+]
 
 const PORT = process.env.PORT || 9000
 
@@ -11,18 +21,19 @@ const server = express()
 const config = {
   entry: server, // or specify port to start io server
   defaultValue,
-  saveTreshold: 2000,
+  saveFrequency: 2000,
   onAuthRequest: async (query, socket) => {
     // some query validation
     return true
   },
   onDocumentLoad: async pathname => {
-    // return initial document ValueJSON by pathnme
+    // request initial document ValueJSON by pathnme
     return defaultValue
   },
-  onDocumentSave: async (pathname, document) => {
+  onDocumentSave: async (pathname, doc) => {
     // save document
+    // console.log('onDocumentSave', pathname, doc)
   }
 }
 
-const connection = new Connection(config)
+const connection = new SocketIOConnection(config)
